+++ fee/engine.go
@@ -140,7 +140,7 @@ func (e *Engine) CalculateForContinuousMode(
 			Owner: aggressor,
 			Amount: &types.FinancialAmount{
 				Asset:  e.asset,
-				Amount: int64(fee.MakerFee),
+				Amount: fee.MakerFee,
 			},
 			Type: types.TransferType_TRANSFER_TYPE_MAKER_FEE_PAY,
 		})
@@ -149,7 +149,7 @@ func (e *Engine) CalculateForContinuousMode(
 			Owner: maker,
 			Amount: &types.FinancialAmount{
 				Asset:  e.asset,
-				Amount: int64(fee.MakerFee),
+				Amount: fee.MakerFee,
 			},
 			Type: types.TransferType_TRANSFER_TYPE_MAKER_FEE_RECEIVE,
 		})
@@ -160,7 +160,7 @@ func (e *Engine) CalculateForContinuousMode(
 		Owner: aggressor,
 		Amount: &types.FinancialAmount{
 			Asset:  e.asset,
-			Amount: int64(totalInfrastructureFeeAmount),
+			Amount: totalInfrastructureFeeAmount,
 		},
 		Type: types.TransferType_TRANSFER_TYPE_INFRASTRUCTURE_FEE_PAY,
 	})
@@ -169,7 +169,7 @@ func (e *Engine) CalculateForContinuousMode(
 		Owner: aggressor,
 		Amount: &types.FinancialAmount{
 			Asset:  e.asset,
-			Amount: int64(totalLiquidityFeeAmount),
+			Amount: totalLiquidityFeeAmount,
 		},
 		Type: types.TransferType_TRANSFER_TYPE_LIQUIDITY_FEE_PAY,
 	})
@@ -378,7 +378,7 @@ func (e *Engine) CalculateFeeForPositionResolution(
 			Owner: goodParty,
 			Amount: &types.FinancialAmount{
 				Asset:  e.asset,
-				Amount: int64(fees.MakerFee),
+				Amount: fees.MakerFee,
 			},
 			Type: types.TransferType_TRANSFER_TYPE_MAKER_FEE_RECEIVE,
 		})
@@ -426,10 +426,10 @@ func (e *Engine) BuildLiquidityFeeDistributionTransfer(shares map[string]float64
 		ft.transfers = append(ft.transfers, &types.Transfer{
 			Owner: key,
 			Amount: &types.FinancialAmount{
-				Amount: int64(cs),
+				Amount: uint64(cs),
 				Asset:  acc.Asset,
 			},
-			MinAmount: int64(cs),
+			MinAmount: uint64(cs),
 			Type:      types.TransferType_TRANSFER_TYPE_LIQUIDITY_FEE_DISTRIBUTE,
 		})
 	}
@@ -438,7 +438,7 @@ func (e *Engine) BuildLiquidityFeeDistributionTransfer(shares map[string]float64
 	last := keys[len(keys)-1]
 	diff := acc.Balance - uint64(floored)
 	ft.totalFeesAmountsPerParty[last] += diff
-	ft.transfers[len(ft.transfers)-1].Amount.Amount += int64(diff)
+	ft.transfers[len(ft.transfers)-1].Amount.Amount += diff
 
 	return ft
 }
@@ -448,9 +448,9 @@ func (e *Engine) BuildLiquidityFeeDistributionTransfer(shares map[string]float64
 func (e *Engine) getPositionResolutionFeesTransfers(
 	party string, share float64, fees *types.Fee,
 ) ([]*types.Transfer, *types.Fee, uint64) {
-	makerFee := int64(math.Ceil(share * float64(fees.MakerFee)))
-	infraFee := int64(math.Ceil(share * float64(fees.InfrastructureFee)))
-	liquiFee := int64(math.Ceil(share * float64(fees.LiquidityFee)))
+	makerFee := uint64(math.Ceil(share * float64(fees.MakerFee)))
+	infraFee := uint64(math.Ceil(share * float64(fees.InfrastructureFee)))
+	liquiFee := uint64(math.Ceil(share * float64(fees.LiquidityFee)))
 
 	return []*types.Transfer{
 			{
@@ -530,7 +530,7 @@ func (e *Engine) getAuctionModeFeeTransfers(infraFee, liquiFee uint64, p string)
 			Owner: p,
 			Amount: &types.FinancialAmount{
 				Asset:  e.asset,
-				Amount: int64(infraFee),
+				Amount: infraFee,
 			},
 			Type: types.TransferType_TRANSFER_TYPE_INFRASTRUCTURE_FEE_PAY,
 		},
@@ -538,7 +538,7 @@ func (e *Engine) getAuctionModeFeeTransfers(infraFee, liquiFee uint64, p string)
 			Owner: p,
 			Amount: &types.FinancialAmount{
 				Asset:  e.asset,
-				Amount: int64(liquiFee),
+				Amount: liquiFee,
 			},
 			Type: types.TransferType_TRANSFER_TYPE_LIQUIDITY_FEE_PAY,
 		},
